schema modularization and schema validator function
