+++ package.json
@@ -77,7 +77,6 @@
     "express-wolox-logger": "^2.0.0",
     "factory-girl": "^5.0.4",
     "jwt-simple": "^0.5.6",
-    "lodash": "^4.17.21",
     "pg": "^8.7.1",
     "sequelize": "^6.9.0",
     "swagger-ui-express": "^4.1.6",
